Update dupeGuru.download.recipe
Ownership of dupeGuru has changed, and the Sparkle Feed no longer exists. It's easy enough to scrape the new webpage https://dupeguru.voltaicideas.net. One could possibly also use GitHubReleaseProvider, but since the macOS version is behind the latest, that might be frail.

The app is not signed, so no point in `CodeSignatureVerifier`. Hence: omitted.